Fix: 이음보드 api 수정

See merge request s10-final/S10P31E203!158

+++ Backend/moass-ws-server/src/main/java/com/moass/ws/repository/BoardRepository.java
@@ -3,5 +3,5 @@ package com.moass.ws.repository;
 import com.moass.ws.entity.Board;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BoardRepository extends JpaRepository<Board, Long> {
+public interface BoardRepository extends JpaRepository<Board, Integer> {
 }
